Update to be able to yarn install with node v12
Updated all packages with their engines locked to 10 to be >= 10.

Updated dependency on old phone-number-privacy module that still
had its engine locked to 10.

Resolved problem with scrypt not building on node > 10.

The dependency on scrypt was as follows:
ethereumjs-wallet -> scrypt.js -> scrypt

Although a pr was raised to fix scrypt
(https://github.com/barrysteyn/node-scrypt/pull/197) the module owner
has lost his npm key and can't publish an updated version.

scrypt.js still depends on scrypt in the latest version, so a resolution
to override the version of ethereumjs-wallet (with a resolution for a
more recent version) seemed like the most concise way of avoiding a
dependency on scrypt.